+++ azure-pipelines.yml
@@ -52,7 +52,7 @@ jobs:
           IsShipped: false
           DropsToRetain: 'CodeAnalysisLogs'
         condition: and(
-          startsWith(variables['Build.SourceBranch'], 'refs/tags/v2.'),
+          startsWith(variables['Build.SourceBranch'], 'refs/tags/v2.0.0-beta'),
           in(variables['Agent.JobStatus'], 'Succeeded', 'SucceededWithIssues'),
           ne(variables['Build.Reason'], 'PullRequest'),
           ne(variables['Build.Reason'], 'Manual')
@@ -84,7 +84,7 @@ jobs:
 
       - task: ComponentGovernanceComponentDetection@0
         condition: and(
-          or(eq(variables['Build.SourceBranch'], 'refs/heads/2.0-preview'), startsWith(variables['Build.SourceBranch'], 'refs/tags/v')),
+          or(eq(variables['Build.SourceBranch'], 'refs/heads/2.0-preview'), startsWith(variables['Build.SourceBranch'], 'refs/tags/v2.0.0-beta')),
           ne(variables['Build.Reason'], 'PullRequest'))
         displayName: 'Component Governance (2.0-preview, release)'
