+++ tools/cli/preRelease.js
@@ -1,14 +1,44 @@
 /* eslint-disable */
 
+const { exec } = require('child_process');
 const fs = require('fs');
 const path = require('path');
-const { buildAndGetIntegrityHash } = require('./utils');
 
-const relativePathToTeamsJsPackageJson = '../../packages/teams-js/package.json';
-const relativePathToTeamsJsReadme = '../../packages/teams-js/README.md';
+const execShellCommand = async cmd => {
+  return new Promise((resolve, reject) => {
+    exec(cmd, { maxBuffer: 1024 * 500 }, (error, stdout, stderr) => {
+      if (error) {
+        reject(error);
+      } else if (stderr) {
+        // Most cli programs output logs to stderr
+        // Breaking errors would be rejected in the prior case
+        console.log(stderr);
+        resolve(stderr);
+      } else {
+        resolve(stdout);
+      }
+    });
+  });
+};
 
-const relativePathToTestAppPackageJson = '../../apps/teams-test-app/package.json';
-const relativePathToTestAppHtml = '../../apps/teams-test-app/index_cdn.html';
+const buildAndGetIntegrityHash = async () => {
+  const relativePathToManifestJson = '../../packages/teams-js/dist/MicrosoftTeams-manifest.json';
+  const absolutePathToManifestJson = path.resolve(__dirname, relativePathToManifestJson);
+
+  console.log('Building @microsoft/teams-js');
+  await execShellCommand('yarn workspace @microsoft/teams-js build');
+
+  if (!fs.existsSync(absolutePathToManifestJson)) {
+    throw `ERROR: MicrosoftTeams-manifest.json at ${absolutePathToManifestJson} was not found.`;
+  }
+  const manifestFile = fs.readFileSync(absolutePathToManifestJson);
+  const manifestJson = JSON.parse(manifestFile);
+  const integrityHash = manifestJson['MicrosoftTeams.min.js']['integrity'];
+  if (!integrityHash) {
+    throw 'ERROR: MicrosoftTeams.min.js integrity hash value was not parsed';
+  }
+  return integrityHash;
+};
 
 const updatePackageJson = (absolutePath, version) => {
   console.log(`Updating ${absolutePath} version to ${version}`);
@@ -40,6 +70,11 @@ const updateVersionAndIntegrity = async (absolutePath, version, integrityHash) =
     process.exit(1);
   }
   try {
+    const relativePathToTeamsJsPackageJson = '../../packages/teams-js/package.json';
+    const relativePathToTeamsJsReadme = '../../packages/teams-js/README.md';
+    const relativePathToTestAppPackageJson = '../../apps/teams-test-app/package.json';
+    const relativePathToTestAppHtml = '../../apps/teams-test-app/index_cdn.html';
+
     const absolutePathTeamsJsPackageJson = path.resolve(__dirname, relativePathToTeamsJsPackageJson);
     const absolutePathTestAppPackageJson = path.resolve(__dirname, relativePathToTestAppPackageJson);
     const absolutePathToTeamsJsReadme = path.resolve(__dirname, relativePathToTeamsJsReadme);
